- small cleanups - introducing checks on time and memory limits - they are still not ideal ...

+++ src/test/resources/Fib.java
@@ -3,7 +3,7 @@ public class Solution {
      * @param n id of fibonacci term to be returned
      * @return  N'th term of Fibonacci sequence
      */
-    public int fib(int n) {
+    public long fib(int n) {
         if (n <= 0) return 0;
         if (n == 1) return 1;
